+++ src/main/java/com/smoky/bassshakertelemetry/client/ClientInit.java
@@ -15,6 +15,8 @@ public final class ClientInit {
         MinecraftForge.EVENT_BUS.register(new TelemetryEventHandler());
         MinecraftForge.EVENT_BUS.register(new SoundHapticsHandler());
         MinecraftForge.EVENT_BUS.register(new GameplayHapticsHandler());
+        MinecraftForge.EVENT_BUS.register(new MovementHapticsHandler());
+        MinecraftForge.EVENT_BUS.register(new MiningSwingHapticsHandler());
 
         ModLoadingContext.get().registerExtensionPoint(
                 ConfigScreenHandler.ConfigScreenFactory.class,